Collections

Collections

The current EquationOfStates contain

EquationOfState
├─ AntonSchmidt
├─ BreenanStacey
├─ FiniteStrainEquationOfState
│  ├─ BirchMurnaghan2nd
│  ├─ BirchMurnaghan3rd
│  ├─ BirchMurnaghan4th
│  ├─ PoirierTarantola2nd
│  ├─ PoirierTarantola3rd
│  └─ PoirierTarantola4th
├─ Murnaghan
├─ PolynomialEquationOfState
└─ Vinet

Types

EquationsOfState.Collections.EquationOfState — Type.

EquationOfState{T,N} <: FieldVector{N,T}

An abstraction of equations of state, where T specifies the elements' type, and N specifies the number of fields.

EquationOfState{T,N} is the abstraction of all equations of state. Subtype it with your customized T and N. It is also a subtype of FieldVector from package StaticArrays.jl.

source

EquationsOfState.Collections.FiniteStrainEquationOfState — Type.

FiniteStrainEquationOfState{T,N} <: EquationOfState{T,N}

An abstraction of finite strain equations of state.

source

EquationsOfState.Collections.Murnaghan — Type.

Murnaghan(v0, b0, bp0, e0=0)

Create a Murnaghan equation of state. The elements' type will be handled automatically.

Arguments

v0: the volume of solid at zero pressure.

b0: the bulk modulus of solid at zero pressure.

bp0: the first-order pressure-derivative bulk modulus of solid at zero pressure.

e0=0: the energy of solid at zero pressure. By default is 0.

source

EquationsOfState.Collections.BirchMurnaghan2nd — Type.

BirchMurnaghan2nd(v0, b0, e0=0)

Create a Birch–Murnaghan 2nd order equation of state. The elements' type will be handled automatically.

Arguments

v0: the volume of solid at zero pressure.

b0: the bulk modulus of solid at zero pressure.

e0=0: the energy of solid at zero pressure. By default is 0.

source

EquationsOfState.Collections.BirchMurnaghan3rd — Type.

BirchMurnaghan3rd(v0, b0, bp0, e0=0)

Create a Birch–Murnaghan 3rd order equation of state. The elements' type will be handled automatically.

Arguments

v0: the volume of solid at zero pressure.

b0: the bulk modulus of solid at zero pressure.

bp0: the first-order pressure-derivative bulk modulus of solid at zero pressure.

e0=0: the energy of solid at zero pressure. By default is 0.

source

EquationsOfState.Collections.BirchMurnaghan4th — Type.

BirchMurnaghan4th(v0, b0, bp0, bpp0, e0=0)

Create a Birch–Murnaghan 4th order equation of state. The elements' type will be handled automatically.

Arguments

v0: the volume of solid at zero pressure.

b0: the bulk modulus of solid at zero pressure.

bp0: the first-order pressure-derivative bulk modulus of solid at zero pressure.

bpp0: the second-order pressure-derivative bulk modulus of solid at zero pressure.

e0=0: the energy of solid at zero pressure. By default is 0.

source

EquationsOfState.Collections.PoirierTarantola2nd — Type.

PoirierTarantola2nd(v0, b0, e0=0)

Create a Poirier–Tarantola order equation of state. The elements' type will be handled automatically.

Arguments

v0: the volume of solid at zero pressure.

b0: the bulk modulus of solid at zero pressure.

e0=0: the energy of solid at zero pressure. By default is 0.

source

EquationsOfState.Collections.PoirierTarantola3rd — Type.

PoirierTarantola3rd(v0, b0, bp0, e0=0)

Create a Poirier–Tarantola 3rd order equation of state. The elements' type will be handled automatically.

Arguments

v0: the volume of solid at zero pressure.

b0: the bulk modulus of solid at zero pressure.

bp0: the first-order pressure-derivative bulk modulus of solid at zero pressure.

e0=0: the energy of solid at zero pressure. By default is 0.

source

EquationsOfState.Collections.PoirierTarantola4th — Type.

PoirierTarantola4th(v0, b0, bp0, bpp0, e0=0)

Create a Birch–Murnaghan 4th order equation of state. The elements' type will be handled automatically.

Arguments

v0: the volume of solid at zero pressure.

b0: the bulk modulus of solid at zero pressure.

bp0: the first-order pressure-derivative bulk modulus of solid at zero pressure.

bpp0: the second-order pressure-derivative bulk modulus of solid at zero pressure.

e0=0: the energy of solid at zero pressure. By default is 0.

source

EquationsOfState.Collections.Vinet — Type.

Vinet(v0, b0, bp0, e0=0)

Create a Vinet equation of state. The elements' type will be handled automatically.

Arguments

v0: the volume of solid at zero pressure.

b0: the bulk modulus of solid at zero pressure.

bp0: the first-order pressure-derivative bulk modulus of solid at zero pressure.

e0=0: the energy of solid at zero pressure. By default is 0.

source

Usage

Construct an EquationOfState

We will use BirchMurnaghan3rd as an example.

BirchMurnaghan3rd can be constructed from scratch:

julia> BirchMurnaghan3rd(1, 2, 3)
4-element BirchMurnaghan3rd{Int64}:
 1
 2
 3
 0

julia> BirchMurnaghan3rd(1, 2, 3, 4)
4-element BirchMurnaghan3rd{Int64}:
 1
 2
 3
 4

julia> BirchMurnaghan3rd(1, 2, 3, 4.0)
4-element BirchMurnaghan3rd{Float64}:
 1.0
 2.0
 3.0
 4.0

It can also be constructed from an existing BirchMurnaghan3rd:

julia> BirchMurnaghan3rd(BirchMurnaghan3rd(1, 2, 3, 4.0), b0=10, e0=5)
4-element BirchMurnaghan3rd{Float64}:
  1.0
 10.0
  3.0
  5.0

julia> BirchMurnaghan3rd(BirchMurnaghan3rd(1, 2, 3, 4.0), Dict(:b0=>10, :e0=>5))
4-element BirchMurnaghan3rd{Float64}:
  1.0
 10.0
  3.0
  5.0

julia> BirchMurnaghan3rd(BirchMurnaghan3rd(1, 2, 3, 4.0), (:b0, 10))
4-element BirchMurnaghan3rd{Float64}:
  1.0
 10.0
  3.0
  4.0

Users can access BirchMurnaghan3rd's element by either "dot notation" or indexing:

julia> b = BirchMurnaghan3rd(1, 2, 3, 4.0)
4-element BirchMurnaghan3rd{Float64}:
 1.0
 2.0
 3.0
 4.0

julia> b.v0
1.0

julia> b[1]
1.0

Calculate energies on an EquationOfState

The $E(V)$ relation of equations of state are listed as below:

Murnaghan:

\[E(V) = E_{0}+K_{0} V_{0}\left[\frac{1}{K_{0}^{\prime}\left(K_{0}^{\prime}-1\right)}\left(\frac{V}{V_{0}}\right)^{1-K_{0}^{\prime}}+\frac{1}{K_{0}^{\prime}} \frac{V}{V_{0}}-\frac{1}{K_{0}^{\prime}-1}\right]\]

BirchMurnaghan2nd:

\[E(V) = E_{0} + \frac{9}{8} B_{0} V_{0} \left(\left( V / V_0 \right)^{-2 / 3}-1\right)^{2}\]

BirchMurnaghan3rd:

\[E(V) = E_{0}+\frac{9}{16} V_{0} B_{0} \frac{\left(x^{2 / 3}-1\right)^{2}}{x^{7 / 3}}\left\{x^{1 / 3}\left(B_{0}^{\prime}-4\right)-x\left(B_{0}^{\prime}-6\right)\right\}\]

where $x = V / V_0$, and $f = \frac{ 1 }{ 2 } \bigg[ \bigg( \frac{ V_0 }{ V } \bigg)^{2/3} - 1 \bigg]$.

BirchMurnaghan4th:

\[E(V) = E_{0}+\frac{3}{8} V_{0} B_{0} f^{2}\left[\left(9 H-63 B_{0}^{\prime}+143\right) f^{2}+12\left(B_{0}^{\prime}-4\right) f+12\right]\]

where $H = B_0 B_0'' + (B_0')^2$.

PoirierTarantola2nd:

\[E(V) = E_{0}+\frac{1}{2} B_{0} V_{0} \ln ^{2} x\]

Public interfaces

EquationsOfState.Collections.calculate — Method.

calculate(EnergyRelation, eos::EquationOfState)

Return a function that can take a volume as a parameter, suitable for batch-applying.

Examples

julia> using EquationsOfState, EquationsOfState.Collections

julia> f = calculate(EnergyRelation, Vinet(1, 2, 3));

julia> map(f, 1:1:10)
10-element Array{Float64,1}:
 0.0
 0.367905230584308
 0.7652477289745814
 1.0516459435179233
 1.2560420090256408
 1.405149833626178
 1.5165867441792136
 1.6017034530570884
 1.6679539823686644
 1.7203642945516917

source

EquationsOfState.Collections.calculate — Method.

calculate(EnergyRelation, eos::Murnaghan, v::Real)

Return the energy of a Murnaghan equation of state on volume v.

source

EquationsOfState.Collections.calculate — Method.

calculate(EnergyRelation, eos::BirchMurnaghan2nd, v::Real)

Return the energy of a BirchMurnaghan2nd equation of state on volume v.

source

EquationsOfState.Collections.calculate — Method.

calculate(EnergyRelation, eos::BirchMurnaghan3rd, v::Real)

Return the energy of a BirchMurnaghan3rd equation of state on volume v.

source

EquationsOfState.Collections.calculate — Method.

calculate(EnergyRelation, eos::BirchMurnaghan4th, v::Real)

Return the energy of a BirchMurnaghan4th equation of state on volume v.

source

EquationsOfState.Collections.calculate — Method.

calculate(PressureRelation, eos::EquationOfState)

Return a function that can take a volume as a parameter, suitable for batch-applying.

Examples

julia> using EquationsOfState, EquationsOfState.Collections

julia> f = calculate(PressureRelation, Vinet(1, 2, 3));

julia> map(f, 1:1:10)
10-element Array{Float64,1}:
  0.0                
 -0.45046308428750254
 -0.3384840350043251 
 -0.24010297221667418
 -0.17314062272722755
 -0.12795492664586872
 -0.09677154467733216
 -0.07468060255179591
 -0.05864401631176751
 -0.04674768462396211

source

EquationsOfState.Collections.calculate — Method.

calculate(BulkModulusRelation, eos::EquationOfState)

Return a function that can take a volume as a parameter, suitable for batch-applying.

Examples

julia> using EquationsOfState, EquationsOfState.Collections

julia> f = calculate(BulkModulusRelation, BirchMurnaghan3rd(1, 2, 3));

julia> map(f, 1:1:10)
10-element Array{Float64,1}:
 2.0                 
 0.9216086833346415  
 0.444903691617472   
 0.2540009203153288  
 0.16193296566524193 
 0.11130584492987289 
 0.08076305569984538 
 0.06103515625       
 0.047609811583958425
 0.03808959181078831 

source